fix: Adjust .github structure for org-wide templates
The GitHub documentation on setting up organization-wide community health files is confusing regarding the structure of the `.github` repository. Specifically, issue and pull request templates need to reside in an additional `.github` directory; that is, `.github/.github/ISSUE_TEMPLATE/<issue-template>.md` and `.github/.github/pull_request_template.md`. This commit adjusts our structure to align with the working implementation.

See the docs for reference: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file#creating-a-repository-for-default-files